rename formatToPartsWithTimezone() to getDateParts() and move
